requirements.txt
# requirements.txt
Flask
pandas
numpy
joblib
matplotlib
google-genai